Update Fractional Limit and Fractional Offset processors to use Float32

+++ src/Processors/FractionalLimitTransform.cpp
@@ -17,8 +17,8 @@ namespace ErrorCodes
 
 FractionalLimitTransform::FractionalLimitTransform(
     SharedHeader header_, 
-    BFloat16 limit_fraction_, 
-    BFloat16 offset_fraction_, 
+    Float32 limit_fraction_, 
+    Float32 offset_fraction_, 
     UInt64 offset_,
     size_t num_streams,
     bool with_ties_,
